move argv to map strigifier to helpers and add unit tests for it, as well as the functionality to display flags with datamask if specified
